requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Werkzeug==3.1.0
SQLAlchemy==2.0.36
email-validator==2.2.0
Pillow==10.2.0
python-dotenv==1.0.1
WTForms==3.1.2
gunicorn==21.2.0
mysqlclient==2.2.1
Flask-Mail==0.9.1
Flask-Babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
PyMySQL==1.1.0
cryptography==41.0.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
